restructuring of record hashing
The record hashing was somewhat hacky till now, like duplicating
functionality and implementing a return-true catch-all for types. The
hashing of RecordTypes is now mapped down to mapping the actual
RecordDeclararion with hashDecl. If the hashing finds a pointer-type in
a struct, we hash the short string representation (e.g., 'struct
next *') instead of recursing into the type.