plugin: ignore EmptyDecl
Too fool ccache, we add a spurious semi-colon in the global space. To
avoid warnings, which could slow down compilation, we have to suppress
the warnings and ignore the empty declaration in the visitor. The
warnings are only visible with -pedantic, which is no problem for the
data we have gathered so far.