Merge local hashes on collision; prevent infinite recursion
On collision of two symbols (e.g. a function declaration in an included header
file and a separate definition), the hashes of those symbols are merged
into a new local hash.

This assumes/requires that the colliding symbols are actually one and the same.

Also fixed recursion problem, where two functions calling each other caused the
script to crash because of infinite recursion in the evaluation.
If the calculation of a symbol's global hash is not yet finished but that symbol
is encountered again, its local hash is used for the calculation instead.