Do not store or reuse dummy hashes
For pointer types inside records, dummy hashes are calculated
as the pointee types might not exist yet.

To prevent different hashes for a record in different files (depending on the
order in which the types are hashed), inside records, no pointer types shall be
stored or reused if already encountered, and the dummy value (a string
representation of the type) shall be used consistently.

T32